Provisions for external Python module development
This commit alters a few things about the way pyarc2 is built in order
to facilitate the development of external python modules that utilise the
low-level libarc2 primitives. In addition to the cdylib used for pyarc2
this crate can also be built as a Rust library. Then the wrapped low-level
Instrument object can be acquired with `PyInstrument.inner` and
`PyInstrument.inner_mut`.