Convert Iterables to ndarrays for functions that require the latter
Many pyarc2 functions require a `PyReadonlyArray<Ix1, type>` or
`PyReadonlyArray<Ix2, type>` which essentially translates to
`ndarray(…, ndims=1, dtype=type)` or `ndarray(…, ndims=2, dtype=type)`.
This requirement means that when calling the function from the Python
side by supplying a list or other Iterable a TypeError would be raised.
This commit wraps the Rust pyarc2 functions at a higher level and will
now convert any compatible Iterable to the proper ndarray. For performance
reasons it is still recommended to pass an ndarray as an argument since
apart from dimension and type checking no copy is performed in that case.